Add files via upload
Added a bit more scripts to the server, most of these were custom scripts. I added a "buff" to be applied to anything who gets their food to 20. I had to add a custom gump that i made that tracks the custom buffs. The goal is to not change the client side and keep everything server side.